Passing the rates as double along with the partitions

+++ src/main/java/org/hps/AbstractAssignor.java
@@ -28,10 +28,10 @@ public abstract class AbstractAssignor implements ConsumerPartitionAssignor {
     public static final class MemberData {
         public final List<TopicPartition> partitions;
         public final Optional<Integer> generation;
-        public final List<Integer> rates;
+        public final List<Double> rates;
 
 
-        public MemberData(List<TopicPartition> partitions, List<Integer> rates,  Optional<Integer> generation) {
+        public MemberData(List<TopicPartition> partitions, List<Double> rates,  Optional<Integer> generation) {
             this.partitions = partitions;
             this.generation = generation;
             this.rates = rates;
